Buff Orcus's spellcasting
As with other demon lord buffs, reflecting two things: he is no longer
covetous, and he is no longer a required boss fight. (A third thing:
under new monster behavior inherited from vanilla, he will expend at
most one wand of death charge on you if he sees you are reflecting or
magic resistant, so you will get the wand mostly full if you defeat
him.)

One option for buffing him was to give him an artifact Wand of Orcus,
which would have been either an artifact wand of death or an artifact
mace or staff like in EvilHack. I didn't really want to take that path,
especially because EvilHack has to go out of its way to make the
artifact unusable by most heroes because it'd be too good otherwise.
Instead, I focused on giving him stronger spellcasting.

Notably, a number of his spells behave like other ones except they
uniquely ignore magic resistance on the hero. This is the sort of
change that might go over poorly, because players may assume that
they're more protected than they actually are, and then get beaten up.
Still, I want to break out of the pattern where Antimagic makes monster
spellcasting more of an annoyance than anything in the late game, and in
order to achieve that something like this has to happen. (I wouldn't
have done this if the touch of death spell were still an instadeath.)

The other big new thing is the "disenchant" spell - this allows him to
hit players where it hurts in their gear, but not as badly as it would
if he could cast destroy armor ignoring magic resistance. (That's still
a possibility later on, if changes are made to that spell so that it
degrades armor before destroying it outright.)