Demon lords instantly destroy Demonbane if given to them as a bribe
As the artifact that is specifically antithetical to their kind, they
want to get rid of it immediately if you're fool enough to just hand it
over.

This has the minor effect that if you hand it over then later fight
and defeat them, you will get back the other bribe items, but will not
get back Demonbane.

Also tweak the messages to use thou pronouns rather than "you", to be
consistent with their parting message (and most other beings who talk to
you), and give an opening message without launching right into the bribe
demands.