Archfiends you do not deal with apply debuffs on the ascension run
This is the final key piece of the Gehennom overhaul: the archfiends are
all present in the game, and are all now optional, but there is now a
penalty for ignoring or fleeing from them. Ideally, this creates sort of
a sliding scale: one extreme is a player who saves themselves a lot of
time by only doing Vlad's and the Wizard's Tower, and can start the
ascension run early, but then will have a very tough time of it for the
rest of the game, and the other extreme is a player who fights or bribes
every archfiend, taking a lot of risks and dangers to fully explore
Gehennom, but then the ascension run is easier.

The median player, though, will probably decide that some archfiends are
worth defeating, and some aren't. The debuffs are partly balanced around
this: for instance, Demogorgon is very hard to reach and defeat - but
his debuff makes you hunger faster, which is an annoyance but given the
abundance of food in the late game, probably not something that most
players will feel like they need to take care of. Or a Barbarian who
doesn't care about spellcasting will not feel the need to defeat Orcus,
etc.

Note that these things are not locked in at the moment you start the
ascension run. You can still detour back and kill or bribe off an
archfiend to get rid of their debuff if you decide you really hate it.
(Until you get to the Planes, I suppose.)

I am not certain these debuffs are all correctly balanced; fortunately,
if they make the game too hard or easy overall, most of them are just
numbers to tweak.

This also adds any active debuffs to enlightenment (and the dumplog),
which a player might find useful to figure out why they are suddenly
able to carry less or are not regenerating as fast.

Another less specific debuff included in this commit is a port of the
SporkHack/UnNetHack/EvilHack behavior where monsters spawn on the
upstairs during the ascension run; here it is implemented strictly in
addition to regular monster spawning (which does not happen on the
upstairs) and the rate at which it happens depends on how many
undealt-with archfiends there are.