Create 'archfiends' xlogfile field
This will be used for at least a Junethack achievement, but also
potentially for other purposes on the scoreboard.

I wanted this to track which fiends you dealt with, but also distinguish
bribed ones from defeated ones since bribing is a lot easier, so those
that were bribed have a $ stuck alongside their name.