Block levelporting within Gehennom
This probably should have been done earlier in the Gehennom overhaul. It
was always a planned part of it, though: especially now that the
Wizard's Tower and Vlad's Tower are in static locations, it would be
fairly easy for a player to levelport directly to them, bypass most of
the Wizard's Tower or Cocytus, and so on.

One consideration was to allow levelporting from Gehennom to outside it,
but that goes against the idea that it's a tough place which is more
challenging: it's cheapened if you can whip out a cursed teleportation
scroll to return to not just a different level, but a safe level on
which no difficult monsters will generate. (I would have more
reservations about this choice if other escape items had been made less
usable, but they haven't. This is only taking away one of those items.)

It is still, however, possible to levelport up from the Valley, to allow
someone who levelported there too early to escape from being pinned
between the Castle and the Valley's undead.