Fix: lregion impossible and unwinnable condition in wizard2 level
The impossible was because the entry uptele teleport region was a single
space that was a ladder, therefore not floor, so when the hero used a
gain level potion to arrive here it would not find a floor space to
place them.

See the diff for a comment on why a hole is necessary; basically, you
could get stuck here forever if you had no way to make your own hole and
hadn't completely solved the wizard3 puzzle.