Bone devils can summon skeletons
This is the differentiation for bone devils: give them a connection to
that bone-ness by giving them thrall skeletons they can create on the
go. They can't do this in D&D, but it makes enough sense to stand on its
own. Skeletons go down easier than the devil will, but they're not
purely chaff either.

The reason this formula accounts for distance to the player is so that
if you enter a level with a bone devil on the other side of the map and
take your time getting over there, it won't be surrounded by a small
army of skeletons already. Initially I considered looping over the whole
level and seeing if the number of skeletons present exceeded some
multiple of the number of bone devils present, but decided that was
probably not worth doing.