Invert gradients in all the levels that used them
I got a pull request merged upstream a while back that inverted
gradients so instead of selecting points away from the center with
increasing probability, select points near the center with high
probability that decreases with distance. However, this meant that the
gradients already existing in xNetHack had to be inverted when that pull
request code made it here.