Update comment in Arc-loca.lua explaining the random boulder traps
Now that recent 3.7 has been merged, boulder traps can have a launch
point defined - but I attempted to implement them with that and had the
boulders rolling off the map and crashing the game. Effectively, without
the ability to specify the launch2 point, which I don't really care to
implement right now, the design as imagined can't be done.  (And
potentially not even if we had that ability -- I'm not sure the
launching code guarantees the boulder will only roll as far as launch2,
or the edge of the map.)