Fix: silver withering-cure message accounting for blindness
If you are blind, you obviously should not be able to see the silver
blacken. Assume corroding produces some sort of warping or texture
change that can be detected by touch, so there is still always a message
(and thus, not having to implement special logic for not knowing the
silver thing you just rubbed has changed at all).