Remove kludge in styxmarsh.lua for nonrotting corpses
Previously these corpses were created with obj.new and then placed with
placeobj, because it was the only way to obtain the object within Lua to
call stop_timer on. One thing added in the recent vanilla merge was that
des.object() now also returns the object and allows the Lua script to do
things with it, so that can be used instead.