Remove hilite_hidden_stairs
I don't believe this was a particularly widely used option, being used
only in the tty windowport, and it was also prone to a number of
annoying bugs such as the red background color spilling over to
adjacent squares. In any case it has been more or less supplanted by
using #terrain.