Fix: multiple explosion messages
At some point vanilla added a "Boom!" to explode() calls. This is
generally fine, but not in the case of exploding doors, because they
already have a KABOOM!! message. Suppress that. (This is not an issue in
vanilla because its trapped doors do not call explode().)

Also added a comment explaining message sequencing for door explosions
after I got most of the way towards reimplementing the vanilla "KABOOM!!
You see a door explode." message then realized the existing message
sequencing already works.