Fix: misc items hitting monsters produced sear message but not damage
This only happened for things that fell into the "misc obj" code path,
so not ordinary weapons. I believe this bug predates the recent merge -
there is a comment in there about how dmgval() already applies the sear
damage so there is no need to do it again, but in fact dmgval() is never
called here, as confirmed by gdb. Fix this by inserting the call to
dmgval().