When listing monster colors, show each entry in that color
This was a cool thing made possible by the ability to color menu options
introduced in a vanilla change to add_menu, so I did it.

In the process, I learned that specifying a color in add_menu overrides
any custom menucolors that might affect it. So if the list of colored
monsters were previously subject to menucolors, it now isn't.