Remove mazes between Medusa and the Castle
This is an indirect means of fixing another problem, which is that
during the vanilla merge I commented out code that created the maze
background for the "default" maze, and in the process made the game
unwinnable if any levels between Medusa and the Castle happened to
decide to generate as mazes, because no maze would be created and
mazexy() would fail and cause a panic.

Instead I wanted to take the opportunity to get rid of full-level mazes
entirely, which is something I have wanted to do for a while now, and
aligns with the philosophy of removing Gehennom mazes - it's simply less
fun to slog through mazes than regular levels. A separate side reason is
that I want to focus on dungeon level (room-and-corridor) generation,
and this enables more of such levels to exist.