When a level fails to generate, create the "Bug Room"
This replaces the previous default behavior of "making a maze" when a
level fails to load. I'd had this idea for a while - a maze is dark and
twisty and stocked with monsters and objects and traps which are often
out of place. I guess it is intended to make a level that feels like
it's still an (illogical) part of the dungeon, but the problem is that
immersion is broken when you get a Lua error and impossible before you
appear in a level that doesn't look like what you were expecting.

Thus, the Bug Room. Very basic, very spartan, contains the necessities
for continuing to traverse the dungeon but nothing else. It should never
get seen by players in a real game, and if you find yourself in it,
something has gone wrong.

This section at the end of makemaz also was intended for making a
randomized maze when makemaz("") was called with no defined prototype
level file for the dungeon - but now that mazes never appear in Dungeons
of Doom levels and the Gehennom branches that have filler levels, this
case now causes an impossible before creating the Bug Room.