Remove the maze room code
I'm a little sad to see this go, because I spent a good deal of time on
making it work, but... it was always a temporary solution, to make
filler maze levels more open and more interesting. But now there are no
more filler mazes, and there won't be any more added, so there's no
point in keeping the code around.

This also deletes some vanilla code that is not used anymore because
there is no filler maze generation, like the selection of a suitable
vibrating square space (already shouldn't have been used assuming
invocation.lua loaded successfully, but would still have been used as a
fallback if it randomly generated a maze).