Fix: pools created by gushing fountain deleted ammo in destroyed traps
Reported by tux3 on github, this is another case where a raw deltrap
left in after the trap ammo overhaul was actually not safe to call,
because delfloortrap permits bear traps and land mines to be destroyed,
and those contain ammo. Change this to place the ammo (at the bottom of
the newly formed pool).

Also noticed a possible unintialized use of otmp in deltrap_with_ammo
itself when deleting a trap with no ammo, so I fixed this.