Use contents funcs in styxmarsh to get rid of reset_coordinate_system
One change I missed in the vanilla merge was this one by paxed: "When
using des.map contents-function the commands following the map are not
relative to it."

I checked all the uses of des.map with a table argument to make sure
none of the xNetHack specific lua code was using a des.map with a
contents function and then assuming things after it, after the des.map
completed, would still be relative to the map. I didn't find any
cases of this, but I did notice that it makes the heavy-handed repeated
resetting of the coordinate system in styxmarsh.lua unnecessary,
provided that the small individual maps are all provided with their own
contents function. So I added those and got rid of the
reset_coordinate_system calls.

At the last minute I noticed that this is also no longer needed in
hellfill.lua since all the prefabs with maps have a contents function,
which therefore means there is no need for reset_coordinate_system to be
a thing at all, so I removed the entire function.