Fuzzer optimization: avoid pacifist safety prompt
Noticed when running the fuzzer that it spends a visible amount of time
on this prompt which is intended to guard human pacifist players against
attacking a monster without meaning to. Turns out any execution where it
manages to go the 100-turn grace period without attacking anything would
result in it getting endlessly prompted to do this, and it won't ever
type in "Yes" to that prompt, so lots and lots of randomness and cycles
are wasted.

Fix this by never giving this prompt if the fuzzer is turned on.