Fix: fixed-dest teleportation traps didn't check if dest was occupied
The point of these teleportation traps with a specific destination is
that they always bring you to that exact spot. Somehow when I wrote the
code for this I failed to consider that the destination may be occupied
by a monster or the player, causing the monster or player to be placed
directly on top of the resident of that space, which caused a crash.
This should now be gracefully handled.

I implemented these non-symmetrically for the player and for monsters -
the player forces a monster on the destination to relocate elsewhere,
while a monster will just fail to teleport - because it best suits the
Wizard teleport maze this way. Players will not be able to engineer a
situation where they get bumped into an adjoining room they shouldn't be
able to access from their current room, nor will they be frustrated by a
sessile monster randomly spawning on the destination of a particular
important teleporter and unable to do anything to get them off it.

While here, I additionally added a guard against creating a
specific-destination teleport trap that sends the triggerer to its own
space, because it doesn't make sense.