Fix: another raw deltrap call that did not handle ammo
In this case, when there is a level that generates stairs after traps
(which happened in hellfill in this case - makes sense, since the
prefabs and any traps they might add should come before stair
placement), it's possible the location chosen for stairs has a trap on
it. In this case, the trap is going away and its ammo should not be
preserved.

I also did move the stair creation in hellfill to just after the prefabs
get placed down, rather than at the very end after monster/object/trap
creation.