Fix: metallivore eating ball/chain and not consuming it
Leading to a panic when it was still considered attached to the hero but
present in the monster's stomach. Unconditionally destroy the ball or
chain if a metallivore eats it. (Another slightly more complex solution
would have been to let them store it in their stomach and just clear
uball/uchain, but this is probably a fringe case anyway.)