Fix: Frontier Town upstairs often didn't get placed
This is due to the change in vanilla that a map with a contents function
now resets the coordinate system after the function ends. In Frontier
Town, the map had a large contents function, followed by a few
directives placing level-wide things, including two levregions that
placed the stairs on either side. The problem is that since the
coordinate system was reset, the area meant to exclude the stairs
from appearing inside the town didn't have the context of where the town
map was anymore.

Fix this by moving the levregion inside the town contents function even
though the stairs aren't actually town contents.