Change how killed-all-archfiends is awarded in the xlogfile
bhaak didn't like the archfiends field and having to do custom parsing
on it for Junethack. So set it up as an achieveX instead.

While here, I added XNH_EXTRA_ACHIEVEMENTS so that adding further
achieveX awards won't eventually potentially cause a buffer overflow
because it only alloted a certain amount of space for N_ACH.