Buff gain energy potions
Note by copperwater: This is intended to make them more reliable as a
method of recovering a good chunk of energy quickly by a spellcaster;
since they're consumable, it's fine for them to be pretty good and a lot
better than ordinary energy regeneration.

Adding a (expensive, top-level) way to alchemize them is from Fourk.