Zombies can't open doors
Note by copperwater: the motivation for this is to add a bit more flavor
for zombies -- depictions in pop culture vary of course, but many have
them as shambling, unthinking beings which lack the ability to do things
like intentionally lift a door latch. This gives a player being pursued
by zombies an extra tactical option to put a door between them (of
course, this was already a tactical option for a lot of other monsters
too.)

Ideally zombies ought to be able to open some doors by simply pushing,
from one direction, if they have no latch or are ajar, but NetHack
doesn't store data like that about doors. So the assumption is that all
doors latch when closed.

(I considered having the condition be _mindless_ and not just zombies
specifically, but other undead which NetHack treats as mindless --
ghouls, mummies, skeletons -- are frequently depicted in other media as
being a little smarter than brainless zombies, enough at least to figure
out a door latch.)

Closes #140
Closes #166