Credit cards may slip through a lock when cursed or hero is fumbling
Note by copperwater: Another slight nerf to credit cards as an unlocking
tool. In the original concept of the idea this was possible to happen
for a non-cursed credit card when the hero wasn't fumbling, but I
figured it should only really punish when there's some negative
circumstance mucking things up.

As implemented this patch had a flat 5% chance if cursed or fumbling; I
modified it back to the original idea which is Dex based. I also noticed
that the new code was inserted at the wrong point, before the "still
busy" check, meaning this check would have triggered every turn of
unlocking, rather than once per the unlock attempt. Moved it so that it
triggers after the unlock attempt resolves, which is not perfect
(ideally it would have a chance of happening at any point during the
process) but good enough.