Fix: Juiblex could be summoned an infinite number of times
This is due to an interaction with the code that prevents him from being
made extinct (if he was, he couldn't split). Fix this by changing the
check for a summoned demon lord to respect its born counter rather than
its extinction status (should be the same thing as long as this only
considers uniques).