Fix: aklyses got caught in paranoid throw "invalid weapon" case
They're a bit unique in this regard, the only typical melee weapon that
is specially designed to be thrown. (Even if throwing a club is unlikely
to be an effective attack, players are accustomed to throwing it, and it
wouldn't be much use if it were rendered invalid for throwing.) Carve
out an exception specifically for them.