Prevent instadeath when Wizard's Tower puzzle is skipped
... and the hero reenters the level from the ladder in the center, then
tries to move out. Without the puzzle being marked as solved, the wall
would slam down and crush the hero when they stepped into the wall gap.

I later realized this should only be possible in wizard mode, so added
an impossible for non-wizard mode appropriately.

I also thought at first it was necessary to fix things if the hero moves
rings around and does not solve the puzzle. Then they will be stuck in
the center, or if they go back upstairs and fall through the hole, stuck
outside the puzzle. Narrowing the criteria such that the puzzle is only
set to solved when the level has just been newly generated fixed that
issue.