Allow monsters to make a superior melee attack over a wand zap
Note by copperwater: This is from EvilHack. It specifically targets the
problem of heavy hitters like Asmodeus wanting to expend their entire
arsenal of offensive items on the player, usually to little or no
effect, before attacking normally. Now, a sufficiently powerful-in-melee
monster who IS in melee range and has offensive items is less likely to
use them.