Fix: articles in dropping-cursed-gold messages
When trying to insert a cursed gold item into a container, the message
started with the (lowercase) object name, sans article -- e.g.  "gold
ring won't leave your person".  Make sure 'The' is included at the start
of the pline when needed.

On the other hand, 'the' was hardcoded into the format string when
failing to drop items with 'd'/'D', so the article was included even if
it didn't belong (e.g. trying to drop cursed Itlachiayaque).  Use the()
there instead.