Fix: hanging when loading Val-loca.lua
This level tries to generate a clear path through the ice on the left
hand side of the screen, which involves retrieving the name of the
terrain type and comparing it against a list of allowed types.  However,
the array of terrain type names had desynced from the actual list of
terrain types, causing it to produce the wrong names so that it would
never actually certify the path as "good to go".  Fix this by
resynchronizing the lists, and add an explicit length to the terrain
type names list based on the enum so that it will produce a compiler
error if more terrains are added in the future without fixing the
terrain name list.