Fix: xlogfile formatting stuck in creature form
"helpless" was being inserted into the 'while' xlogfile field when the
hero died in creature form, even if gm.multi was zero (i.e. the hero
wasn't helpless).  In that case the connecting 'and' of "helpless and
stuck in creature form" wouldn't be included, either, making a while
reason "helplessstuck in creature form". Construct the 'while' field a
bit differently (building it piece by piece with Strcpy/Strcat/Sprintf)
so "helpless" isn't necessarily included and the formatting will work
for all the possible combinations of helpless/stuck.

A similar issue was that if polyinit mode was on, the 'while' reason
would always be "helpless", because the condition to insert a 'while'
reason at all just checked 'stuck' without '&& !Polyinit_mode'.  I moved
the Polyinit_mode check into the definition of 'stuck' since it's (now)
never used without checking Polyinit_mode anyway.