Fix: "leather armor" shim made all wishes produce light armor
The strncmpi was backwards, so any wish string that did NOT match
"leather armor" resolved to a light armor.

Just negating it didn't do the job, though, because "leather" is
stripped when object materials are read. Fortunately that is stored in
d->material though, so check for that and match the string against
"armor".

(Note that if there's ever a piece of armor whose name literally starts
with "armor", a wish for a "leather armor of foo" will still go straight
to the light armor.)