Don't mention material in #call prompt
When #calling a particular type of item, omit information about the
specific inventory item's material (since you're naming the type, not
that particular item) if it's possible to do so.  So prompt "call a
crested helmet" instead of "call a silver crested helmet".  Material is
still mentioned when naming a particular item.

This makes the previous commit somewhat redundant but not entirely,
since objects which always include the material name don't have the
material changed.  (Though almost none of those types are callable
anyway.)