Bump PREFIX size
It needs to be larger now that "silver-scaled" (etc) is a valid prefix
term (in addition to all the others).  A particularly long prefix (such
as the example I added in the comment) can blast past the end of
prefix[] in doname_base without this change.

The comment with the example of a prefix, and the "(56)" comment, was a
little confusing -- I don't know if that was meant to be the maximum
possible prefix length (which is how I read it) or an explanation of why
PREFIX _used to be_ 56 characters long (which is how bhaak read it), but
I changed it to a longer example of an xnh prefix, one that should be
close to the current maximum length.