Fix: Orcus lair wall gap "fallback"
A player recently got an Orcus level where the wall had no gap, so the
level was completely inaccessible.  Looks to me like this was because
the intended fallback position for failure to create any random gaps was
replacing the vertical wall with a horizontal wall, rather than with a
gap.