Fix github mac/windows builds by fixing Lua fetch target to HTTPS
As extra reinforcement, pass -L to curl.

Upstream NetHack has made a bunch of changes to Makefile.top to fix
fetch-lua but merging them all in is too much work right now; when
vanilla is next merged the change from this commit will be dropped when
it conflicts with theirs.