Fix: Mjollnir "doesn't seem designed to be thrown" for Valkyrie
The check accounted for aklyses being designed to be thrown since they
auto-return, but Mjollnir was not counted, meaning Valkyries trying to
throw it would have to say yes to the prompt every time. Replace it with
the AutoReturn macro which accounts for all auto-returning weapons.

Note that because of how AutoReturn behaves, Mjollnir is only treated as
an auto-returning weapon *if the player is a Valkyrie*. This means that
other roles trying to throw it, even if wearing gauntlets of power and
physically able to, will still get the "doesn't seem designed to be
thrown" prompt. Perhaps because they are not as familiar with the lore
of Mjollnir as a Valkyrie would be, the idea of throwing such a weapon
seems strange.

Fixes #208