Escape from controlled polymorph prompt no longer causes random poly
This was one of the remaining "a single typo causes a major gameplay
disruption" behaviors (i.e. turning into something large and breaking
armor, or into a dragon). Since this prompt only appears when the
polymorph is controlled, there's no reason to punish this in particular,
so I changed it to instead inform the player how they can explicitly
request a random polymorph, then prompt again.

However, I did not touch the "That's enough tries! Have a random
polymorph!" behavior if the loop is triggered enough times (currently
5).