Characters who don't like baths may resist fountain's urging
Adapted from comments on PR #175. Rather than a YAFM for the
"bath-unfriendly" groups (cavemen, orcs, and barbarians) that was
cosmetic, this makes it so they have a Wisdom-based chance to avoid the
urge to take a bath. This can be good (without taking a bath, you cannot
lose gold in the fountain) or bad (if you looted the fountain by finding
coins or gems in it, losing gold in the fountain is the only way to
unloot the fountain).