Message tweak for faulty amulet of life saving
Remove the second reference to "your medallion" from "Your medallion
begins to glow! But... the chain on your medallion breaks..." as being
redundant.

While here, I also strengthened the checks in faulty_lifesaver() - null
object case didn't return, so it would have caused a segfault after
printing an impossible, and I felt like it needed an explicit check that
the object is actually an amulet of life saving, since that might (but
isn't likely to) change in the future.