Remove has_ceiling check for scroll of earth
Vanilla added code designed to accomodate for scroll of earth when read
without a ceiling overhead last year - "an avalanche of boulders
materializes above you". In vanilla, levels aren't treated as outdoors
in the same way as xNetHack, so it worked okay in the hardcoded case of
quest levels, but that code wouldn't actually trigger in xNetHack
outdoors levels. Fortunately, all that was required to fix it was
removing the has_ceiling check on scrolls of earth.