Fix: Yeenoghu wand of wishing weirdnesses
Yeenoghu was the only archfiend who carried a wand of wishing upon
generation rather than having the wand somewhere on his level. The
consequence of this was that if he were summoned early by another
archfiend or by same-race sacrifice, he wouldn't have that wand (because
the wand was only given to him when he generated in his level in Lua)
and loading his level would not generate him.

There was probably a way around this by checking to see if a monster
indeed existed at his space after creating him, or making des.monster()
capable of returning a value if the monster were actually created, but
the simpler fix for right now was to put a chest which is always on his
level, which maybe contains the wand and maybe doesn't, like all the
other archfiends.

Also, there was still a debug pline left in the lua file which printed
when Yeenoghu generated with a wand, so removed that.