Zombies are immune to being scared by any source
The initial idea for this was "all mindless undead are immune to active
scare effects, but not the scroll of scare monster". The roadblock I ran
into was that onscary() is really not well set up for that and you'd
have to hack several special cases into otherwise clean conditions to
make it work. Perhaps at some point it could be refactored into a
(monster -> bitmask of things they would be scared by) function, but at
the moment it's easiest to simply add more cases to the "completely
immune to fear" conditional.

To compensate for making the fear immunity broader than originally, I
narrowed the scope to only zombies, which dovetails nicely with the
feature from several commits ago where they can't open doors - other
"mindless" undead may still possess enough instinct or trace of mental
faculties to open doors and avoid scary things, but zombies are dumb as
a rock and don't.

Closes #113