Make martial arts users immune to leg damage from bad kicks
Characters who have at least basic training in proper form for making
kicks are unlikely to fail to apply that training even if they decide to
kick a wall or heavy object. The kick won't necessarily accomplish
anything, but it will be properly executed. This removes the HP and
wounded leg damage in the "Ouch! That hurts!" kick failure case for all
characters with Basic martial arts (i.e. all monks and samurai) and
replaces it with an "ineffective" message. The kick still makes noise to
wake up monsters and still subjects the hero to Newton's 3rd Law.

Closes #151