Fix: two cases of Bifrost not being restored correctly
First bug is issue #212: polymorphing the nemesis and then killing it
meant the killed_nemesis flag never got set, so level updates based on
that flag never triggered. The fix for this is to track the nemesis's
monster ID similar to how it works for the quest leader. This came with
a few tweaks:
- The nemesis's corpse will keep its monster traits.
- The nemesis cannot be tamed.
- You get the big alignment record bonus for killing them.
- They will still talk in their polymorphed form if their form is
  capable of speech.

The second bug involved if the player managed to kill Surtur before
loading Val-loca. This was because of an oversight when the
visited_after_event flag was added - it was not added in
clear_level_structures() and therefore the locate level was registering
as having already processed the post-Surtur-kill change.

Closes #212