Add "random spotlights" lighting variation to partly dark Bigroom
This creates a new variant within the bigrm-2 variant, which are all
about different patterns of darkness on the map. In this one, random
circular regions are lit up and the rest of the room is dark. Had to go
out of my way a bit to implement it so that the lit areas wouldn't
overlap with each other and create misshapen blobs instead of nice
ellipses. I suppose I could have gotten the same effect with less effort
by adding fixed circles of light that are always in the same position,
but it wouldn't have been as good.

Also, this removes the possibility for bigrm-2 to choose no regions of
darkness at all (which results in a room identical to bigrm-1 with none
of its terrain variations, and you already get that from bigrm-1 25% of
the time).

Closes #132