Fix: archfiends generated in bones weren't tracked properly
From a github report, Geryon was loaded in a bones file, and
subsequently killing him did not get rid of his debuff during the
ascension run. The underlying cause is that the num_in_dgn counter did
not tick up when an archfiend was loaded from a bones file, which
subsequently caused the counter to underflow to UINT_MAX when they were
killed. Since that number is greater than 0, fiend_adversity concluded
they were still active in the dungeon.

Fix that problem, and also add a bounds check to guard against the
underflow by printing an impossible if for some reason num_in_dgn
attempts to decrement when already at 0.

Fixes #213